requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipykernel==6.29.2
ipython==8.22.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
keyboard==0.13.5
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
mysqlclient==2.2.6
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.1
notebook_shim==0.2.4
numpy==2.1.2
opencv-python==4.10.0.84
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pillow==11.0.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.10.1
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scipy==1.14.1
Send2Trash==1.8.2
setuptools==75.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.3
terminado==0.18.0
tinycss2==1.2.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
whitenoise==6.8.2
